apps/api-fastapi/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
asyncpg==0.30.0
python-dotenv==1.2.1
